Move mapping_params to aln.hpp
